refactor: move Route to its own file
Moves the `Route` enum to its own file so that it can be imported
without the `getRoute` store which depends on Svelte. This way the enum
can be used in Playwright tests.
